Moved server.auth into commonlib
