+++ openquake/commonlib/auth.py
